requirements.txt
httpx==0.28.1
joblib==1.5.2
matplotlib==3.10.7
numpy==2.3.4
openai==2.7.2
pandas==2.3.3
python-dotenv==1.2.1
scikit_learn==1.7.2
seaborn==0.13.2
streamlit==1.51.0
tqdm==4.67.1
PyMuPDF==1.26.6